requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.1
google-generativeai==0.8.3
Jinja2==3.1.4
lxml==5.3.0
fpdf2==2.8.1
gTTS==2.5.4
eng_to_ipa==0.0.2
deep-translator==1.11.4
